refactor: split views.py into domain-focused modules
Split the 774-line views.py into 4 focused modules:

- auth_views.py: TokenObtainPairView, CreateUserView, GetLoggedInUserView,
  AllUsersView, DebugConfigView
- note_views.py: Notes, NoteDetail
- chat_views.py: Chats, UserChats, ChatDetail
- chat_operations_views.py: ChatSendMessageView, ChatGetHelpView, ChatGradeView

The views/__init__.py re-exports all views for backward compatibility,
so existing imports like `from api.views import Chats` continue to work.

Also updated ruff.toml to apply same per-file-ignores to views package.

All 142 tests pass with 82% coverage.

Closes #68
